--story=1005575 --user=龙儒锋 巡检定义 https://www.tapd.cn/54247054/s/1192004

+++ src/main/java/codedriver/module/inspect/api/InspectCombopSaveApi.java
@@ -51,7 +51,7 @@ public class InspectCombopSaveApi extends PrivateApiComponentBase {
         for (int i = 0; i < inspectCombopList.size(); i++) {
             JSONObject o = JSONObject.parseObject(String.valueOf(inspectCombopList.get(i)));
             String name = String.valueOf(o.get("name"));
-            Integer combopId = (Integer) o.get("combop_id");
+            Long combopId = (Long) o.get("combop_id");
             Document doc = new Document();
             Document updateDoc = new Document();
             Document setDocument = new Document();
